Almost final version
To send a single mail for all the reporting, AsyncTasks have been gathered together in a single one (maybe with more time, we could arrange this).

Furthermore, before sending mails, we check that we do have an internet connection to avoid potential crashing and/or mails stuck in sendbox.

TOFIX:
- outgoing sms and calls still need to examined