Final version
This is the latest version of the project, which is quite complete (documenting the code should be done, but the names of classes and methods used in the program are clear enough for someone who discovers the app to understand very easily).

The reporting of outgoing sms has been abandoned (it seems not to be possible to catch them); otherwise everything should work properly.